+++ example/info_timing.dart
@@ -6,13 +6,14 @@ Future<bool> isDart2AsyncTiming() async {
     // Create an async function
     // in dart1 the first line won't be executed directly
     // in dart2 it should
-    method() async {
+    Future method() async {
       if (_isDart2AsyncTiming == null) {
         _isDart2AsyncTiming = true;
       }
     }
 
     // Calling the async function not waiting for it
+    // ignore: unawaited_futures
     method();
     if (_isDart2AsyncTiming == null) {
       _isDart2AsyncTiming = false;
@@ -21,6 +22,6 @@ Future<bool> isDart2AsyncTiming() async {
   return _isDart2AsyncTiming;
 }
 
-main() async {
+Future main() async {
   print('isDart2AsyncTiming ${await isDart2AsyncTiming()}');
 }