+++ example/synchronized_example.dart
@@ -3,7 +3,7 @@ import 'dart:io';
 import 'package:synchronized/synchronized.dart';
 
 Future writeSlow(int value) async {
-  new Future.delayed(new Duration(milliseconds: 1));
+  await Future.delayed(Duration(milliseconds: 1));
   stdout.write(value);
 }
 
@@ -21,27 +21,31 @@ class Demo {
   Future test1() async {
     stdout.writeln("not synchronized");
     //await Future.wait([write1234(), write1234()]);
+    // ignore: unawaited_futures
     write1234();
+    // ignore: unawaited_futures
     write1234();
 
-    await new Future.delayed(new Duration(milliseconds: 50));
+    await Future.delayed(Duration(milliseconds: 50));
     stdout.writeln();
   }
 
   Future test2() async {
     stdout.writeln("synchronized");
 
-    var lock = new Lock();
+    var lock = Lock();
+    // ignore: unawaited_futures
     lock.synchronized(write1234);
+    // ignore: unawaited_futures
     lock.synchronized(write1234);
 
-    await new Future.delayed(new Duration(milliseconds: 50));
+    await Future.delayed(Duration(milliseconds: 50));
 
     stdout.writeln();
   }
 
   Future readme1() async {
-    var lock = new Lock();
+    var lock = Lock();
 
     // ...
     await lock.synchronized(() async {
@@ -50,16 +54,16 @@ class Demo {
   }
 
   Future readme2() async {
-    var lock = new Lock();
+    var lock = Lock();
     if (!lock.locked) {
-      lock.synchronized(() async {
+      await lock.synchronized(() async {
         // do some stuff
       });
     }
   }
 
   Future readme3() async {
-    var lock = new Lock();
+    var lock = Lock();
     int value = await lock.synchronized(() {
       return 1;
     });
@@ -67,8 +71,8 @@ class Demo {
   }
 }
 
-main() async {
-  var demo = new Demo();
+Future main() async {
+  var demo = Demo();
 
   await demo.test1();
   await demo.test2();