+++ test/issues_test.dart
@@ -5,17 +5,15 @@
 import 'dart:async';
 
 import 'package:synchronized/src/utils.dart';
-import 'package:synchronized/synchronized.dart' hide SynchronizedLock;
+import 'package:synchronized/synchronized.dart';
 import 'package:test/test.dart';
-
-import 'test_common.dart';
 
 void main() {
   group('issues', () {
     // https://github.com/tekartik/synchronized.dart/issues/1
     test('issue_1', () async {
       var value = '';
-      var lock = new Lock(reentrant: true);
+      var lock = Lock(reentrant: true);
 
       Future outer1 = lock.synchronized(() async {
         expect(value, equals(''));
@@ -46,10 +44,6 @@ void main() {
       await Future.wait([outer1, outer2, outer3]);
 
       expect(value, equals('outer3'));
-    });
-
-    test('timing', () async {
-      print("isDart2AsyncTiming ${await isDart2AsyncTiming()}");
     });
   });
 }