+++ tool/analyze_fmt_and_test.sh
@@ -3,8 +3,8 @@
 # Fast fail the script on failures.
 set -xe
 
-dartfmt -w example lib test
-dartanalyzer --fatal-warnings example lib test
+dartfmt --fix -w . example lib test
+dartanalyzer --fatal-warnings --fatal-infos example lib test
 
 pub run test -p vm
 pub run build_runner test